server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==2.3.7
# Data processing (optional - only needed for CSV import scripts)
pandas>=2.2.0  # Updated for Python 3.13 compatibility
numpy>=1.26.0  # Updated for Python 3.13 compatibility
scikit-learn>=1.4.0  # Updated for Python 3.13 compatibility
reportlab==4.0.7
openpyxl==3.1.2
# Testing
pytest>=7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
# PostgreSQL dependencies
psycopg2-binary>=2.9.9  # Updated for Python 3.13 compatibility
SQLAlchemy==2.0.23
